Move BuildOutputCleanupRegistry to :execution
