+++ subprojects/core/src/main/java/org/gradle/internal/cleanup/DefaultBuildOutputCleanupRegistry.java
@@ -19,6 +19,7 @@ package org.gradle.internal.cleanup;
 import com.google.common.collect.Sets;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.file.FileCollectionFactory;
+import org.gradle.internal.execution.workspace.BuildOutputCleanupRegistry;
 
 import java.io.File;
 import java.util.LinkedHashSet;
